+++ test/engine/engine.ts
@@ -42,7 +42,7 @@ for (const fixture of fixtures) {
     nativeIo,
     pluginDisklet: fakeIo.disklet
   }
-  const factory = edgeCorePlugins[fixture.pluginName as 'monero']
+  const factory = edgeCorePlugins[fixture.pluginName as 'beldex']
   const plugin: EdgeCurrencyPlugin = factory(opts)
 
   const emitter = new EventEmitter()