+++ CHANGELOG.md
@@ -22,7 +22,7 @@
 
 # 1.1.0 (2023-08-11)
 
-- added: Support custom servers via `enableCustomServers` and `moneroLightwalletServer` user settings.
+- added: Support custom servers via `enableCustomServers` and `beldexLightwalletServer` user settings.
 
 # 1.0.0 (2023-03-28)